agent/requirements.txt
uagents
google-genai
python-dotenv
fastapi
uvicorn